+++ framework/old/build.sbt
@@ -3,10 +3,9 @@ libraryDependencies ++= Seq(
 )
 
 val slickVersion = "3.2.0-M1"
-val scalazVersion = "7.3.0-M3"
+val scalazVersion = "7.3.0-M5"
 val circeVersion = "0.5.1"
-val poiVersion = "3.15-beta2"
-
+//val poiVersion = "3.15"
 libraryDependencies ++= Seq(
   "com.typesafe.slick" %% "slick" % slickVersion,
   "com.typesafe.slick" %% "slick-hikaricp" % slickVersion exclude("com.zaxxer", "HikariCP-java6"),
@@ -23,8 +22,7 @@ libraryDependencies ++= Seq(
   "io.circe" %% "circe-generic",
   "io.circe" %% "circe-parser"
 ).map(_ % circeVersion)
-
-libraryDependencies ++= Seq(
+/*libraryDependencies ++= Seq(
   //poi
   "org.apache.poi" % "poi" % poiVersion exclude("stax", "stax-api"),
   "org.apache.poi" % "poi-ooxml" % poiVersion exclude("stax", "stax-api"),
@@ -33,4 +31,4 @@ libraryDependencies ++= Seq(
   //joda-time
   "joda-time" % "joda-time" % "2.9.4",
   "org.joda" % "joda-convert" % "1.8.1"
-)
+)*/
