+++ framework/src/main/scala/net/scalax/fsn/mix/helpers/CRUDHelpers.scala
